Use load_only rather than many defer calls
The point of defer_everything_but was to effectively do
what load_only does, but probably back when load_only
was not an option yet.
This just gets rid of it, and uses load_only instead. Since load_only
cannot work on relationships, just columns, we filter
relationships out of the column set before using it.